refactor(e2e): move getSlidersRatingSelector to shared module
