+++ cypress/integration/cbl/progress/overrides.js
@@ -1,3 +1,5 @@
+const { getSlidersRatingSelector } = require('../../../support/ratings.js')
+
 describe('CBL / Progress / Overrides', () => {
 
     // load sample database before tests
@@ -342,24 +344,5 @@ describe('CBL / Progress / Overrides', () => {
             });
 
         cy.wait('@getDemonstrationSkills');
-    };
-
-    // Todo: this function from teacher-submit.js... potential for unification
-    function getSlidersRatingSelector(sliders) {
-        const firstSlider = sliders[0];
-        const { minRating, maxRating } = firstSlider.getConfig();
-        const visibleRatings = Array.from({length: maxRating - minRating + 1 }, (_, i) => i + minRating)
-        const totalThumbs = visibleRatings.length;
-        const ratingWidth = firstSlider.innerEl.getWidth() / totalThumbs;
-
-        return (slider, rating) => {
-            const selectedRatingPos = visibleRatings.indexOf(rating);
-            cy.wrap(slider.innerEl.dom)
-                .click({
-                    x: ratingWidth * (selectedRatingPos + 1),
-                    y: 5,
-                    force: true
-                });
-        };
     };
 });
