+++ cypress/integration/cbl/progress/teacher-submit.js
@@ -1,3 +1,5 @@
+const { getSlidersRatingSelector } = require('../../../support/ratings.js')
+
 describe('CBL / Progress / Teacher Submit', () => {
 
     // load sample database before tests
@@ -211,21 +213,3 @@ describe('CBL / Progress / Teacher Submit', () => {
         });
     });
 });
-
-function getSlidersRatingSelector(sliders) {
-    const firstSlider = sliders[0];
-    const { minRating, maxRating } = firstSlider.getConfig();
-    const visibleRatings = Array.from({length: maxRating - minRating + 1 }, (_, i) => i + minRating)
-    const totalThumbs = visibleRatings.length;
-    const ratingWidth = firstSlider.innerEl.getWidth() / totalThumbs;
-
-    return (slider, rating) => {
-        const selectedRatingPos = visibleRatings.indexOf(rating);
-        cy.wrap(slider.innerEl.dom)
-            .click({
-                x: ratingWidth * (selectedRatingPos + 1),
-                y: 5,
-                force: true
-            });
-    };
-}
